refactor(dash): rename proped-component.ts -> typed-component.ts
