+++ packages/dash/src/lib/common/screen/screen.ts
@@ -1,6 +1,6 @@
 import { get, writable } from 'svelte/store';
 
-import type { TypedComponent } from '$lib/common/component/proped-component.js';
+import type { TypedComponent } from '$lib/common/typed-component.js';
 
 
 export type ScreenHandle = {
