+++ packages/dash/src/lib/common/typed-component.ts
