+++ packages/dash/src/lib/main/page/page.ts
@@ -1,6 +1,6 @@
 import { writable, type Writable } from 'svelte/store';
 
-import type { TypedComponent } from '$lib/common/component/proped-component.js';
+import type { TypedComponent } from '$lib/common/typed-component.js';
 
 export interface Page<T extends Record<string, never> = Record<string, never>> {
     name: string;
