+++ packages/dash/src/lib/main/settings.ts
@@ -1,6 +1,6 @@
 import { writable } from 'svelte/store';
 
-import type { PropedComponent } from '$lib/common/component/proped-component.js';
+import type { PropedComponent } from '$lib/common/typed-component.js';
 import { LOCALES } from '$lib/i18n/i18n.js';
 import Checkbox from './settings/CheckboxField.svelte';
 import Combobox from './settings/ComboboxField.svelte';
